oxenstored: Fix releaseDomain events not being sent on domains shutting down
5e559b4 changed the code in xenstored.ml from this:

    if deaddom <> [] || notify then
      Connections.fire_spec_watches (Store.get_root store) cons
        Store.Path.release_domain

to this:

    if notify then
      List.iter
        (Connections.fire_spec_watches (Store.get_root store) cons
           Store.Path.release_domain
        )
        deaddom

which meant that it would not send any releaseDomain events in case of a
domain shutting down (which is tracked by a variable confusingly named "notify").
Turn this into a list of domains that have shutdown instead, sending
events for each.

This should hopefully fix the issues that required reverting this work (45c16ec)

Signed-off-by: Andrii Sultanov <andriy.sultanov@vates.tech>