oxenstored: Prepare xenstored.ml to be called in the unit tests
Avoid daemonizing and doing some other things that are impossible in the
testing environment outside of xen.

Change Xenstored.main to return its loop body as a function so that unit tests
can step over it one iteration at a time. When not under testing, the
main loop is called as before.

Provide a mock paths.ml and change the mock unix_activations_minimal to
always return the eventchn.domexc port as pending so that Xenstored
checks for dead domains.

Signed-off-by: Andrii Sultanov <andriy.sultanov@vates.tech>